feat: add AuthGuard + profile layout for auth redirect (TRA-439)
Create reusable AuthGuard client component that redirects
unauthenticated users to /login?next=<path>. Wrap profile
routes in new layout.tsx so /profile and /profile/settings
are protected without per-page auth checks.
