+++ apps/web/app/(main)/profile/layout.tsx
@@ -1,0 +1,7 @@
+'use client'
+
+import { AuthGuard } from '@/components/auth/AuthGuard'
+
+export default function ProfileLayout({ children }: { children: React.ReactNode }) {
+  return <AuthGuard>{children}</AuthGuard>
+}
