+++ apps/web/app/(main)/profile/settings/page.tsx
@@ -3,7 +3,7 @@
 import { useState, useEffect, useCallback, useRef } from 'react';
 import { useRouter } from 'next/navigation';
 import { motion, AnimatePresence } from 'motion/react';
-import { User, Compass, Bell, Settings, CheckCircle, Plane, Hotel, Loader2, AlertCircle, Plus, Eye, Check, Shield, Smartphone, LogOut, Star, HelpCircle, MessageSquare, Trash2, X, ChevronDown, Pencil, ClipboardList, Activity, Zap, ArrowLeft } from 'lucide-react';
+import { User, Compass, Bell, Settings, CheckCircle, Plane, Hotel, Loader2, Plus, Eye, Check, Shield, Smartphone, LogOut, Star, HelpCircle, MessageSquare, Trash2, X, ChevronDown, Pencil, ClipboardList, Activity, Zap, ArrowLeft } from 'lucide-react';
 import { toast } from 'sonner';
 import { useQueryClient } from '@tanstack/react-query';
 import { AvatarUpload } from '@/components/AvatarUpload';
@@ -299,21 +299,9 @@ export default function ProfileSettings() {
   // ─── Load Profile Data ───────────────────────────────
   useEffect(() => {
     async function loadProfile() {
-      // Wait for auth to finish loading
-      if (authLoading) {
-        return;
-      }
-
       try {
         setIsLoading(true);
         setError(null);
-
-        // Check if user is authenticated
-        if (!user || !session) {
-          setError('Sign In to View');
-          setIsLoading(false);
-          return;
-        }
 
         // Check if Supabase is configured
         if (!supabase) {
@@ -673,43 +661,6 @@ export default function ProfileSettings() {
           </div>
         </div>
       </div>
-    );
-  }
-
-  if (error && !user) {
-    return (
-      <div className="min-h-screen bg-gradient-to-br from-[#f8fafc] to-[#e0f2fe]">
-        <div className="flex items-center justify-center p-6 pt-24">
-          <div className="max-w-md w-full text-center bg-white rounded-3xl shadow-2xl p-10">
-          {/* Travyl Logo */}
-          <div className="flex items-center justify-center gap-3 mb-8">
-            <span className="text-3xl font-black text-[#1e3a5f] tracking-wider">TRAVYL</span>
-            <svg
-              viewBox="0 0 64 64"
-              className="w-10 h-10"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path d="M60 10 L20 36 L6 34 Z" fill="#ffffff" stroke="#1e3a5f" strokeWidth="2"/>
-              <path d="M48 48 L30 40 L26 38 L60 10 Z" fill="#ffffff" stroke="#1e3a5f" strokeWidth="2"/>
-              <path d="M52 16 L26 38 L24 50 L20 36 Z" fill="#ffffff" stroke="#1e3a5f" strokeWidth="2"/>
-            </svg>
-          </div>
-
-          <div className="w-20 h-20 bg-red-100 rounded-full flex items-center justify-center mx-auto mb-6">
-            <AlertCircle size={40} className="text-red-500" />
-          </div>
-          <h2 className="text-2xl font-bold text-[#1e3a5f] mb-1">Access Denied</h2>
-          <h3 className="text-2xl font-bold text-[#1e3a5f] mb-6">{error}</h3>
-          <a
-            href="/login"
-            className="inline-flex items-center gap-2 px-8 py-3 bg-[#1e3a5f] text-white rounded-xl hover:bg-[#2a4a6f] transition-all font-bold shadow-lg"
-          >
-            Sign In
-          </a>
-        </div>
-      </div>
-    </div>
     );
   }
